+++ packages/liveblocks-react/src/room.tsx
@@ -102,9 +102,9 @@ import type {
   FeedMessagesAsyncSuccess,
   FeedsAsyncResult,
   FeedsAsyncSuccess,
-  HistoryVersionDataAsyncResult,
   HistoryVersionsAsyncResult,
   HistoryVersionsAsyncSuccess,
+  HistoryVersionYjsDataAsyncResult,
   MutationContext,
   OmitFirstArg,
   RoomContextBundle,
@@ -3146,11 +3146,11 @@ function useRoomSubscriptionSettingsSuspense(): [
 /**
  * @internal
  */
-function useHistoryVersionData_withRoomContext(
+function useHistoryVersionYjsData_withRoomContext(
   RoomContext: Context<OpaqueRoom | null>,
   versionId: string
-): HistoryVersionDataAsyncResult {
-  const [state, setState] = useState<HistoryVersionDataAsyncResult>({
+): HistoryVersionYjsDataAsyncResult {
+  const [state, setState] = useState<HistoryVersionYjsDataAsyncResult>({
     isLoading: true,
   });
   const room = useRoom_withRoomContext(RoomContext);
@@ -3158,7 +3158,7 @@ function useHistoryVersionData_withRoomContext(
     setState({ isLoading: true });
     const load = async () => {
       try {
-        const response = await room[kInternal].getTextVersion(versionId);
+        const response = await room[kInternal].getYjsHistoryVersion(versionId);
         const buffer = await response.arrayBuffer();
         const data = new Uint8Array(buffer);
         setState({
@@ -3183,15 +3183,29 @@ function useHistoryVersionData_withRoomContext(
 }
 
 /**
- * Returns the version data bianry for a given version
+ * @deprecated Use `useHistoryVersionYjsData(versionId)` instead.
+ *
+ * Returns the version data binary for a given version
  *
  * @example
  * const {data} = useHistoryVersionData(versionId);
  */
 function useHistoryVersionData(
   versionId: string
-): HistoryVersionDataAsyncResult {
-  return useHistoryVersionData_withRoomContext(GlobalRoomContext, versionId);
+): HistoryVersionYjsDataAsyncResult {
+  return useHistoryVersionYjsData_withRoomContext(GlobalRoomContext, versionId);
+}
+
+/**
+ * Returns the Yjs data for a given version of the room.
+ *
+ * @example
+ * const { data, isLoading, error } = useHistoryVersionYjsData(versionId);
+ */
+function useHistoryVersionYjsData(
+  versionId: string
+): HistoryVersionYjsDataAsyncResult {
+  return useHistoryVersionYjsData_withRoomContext(GlobalRoomContext, versionId);
 }
 
 /**
@@ -4125,10 +4139,10 @@ export function createRoomContext<
     return useHistoryVersionsSuspense_withRoomContext(BoundRoomContext);
   }
 
-  function useHistoryVersionData_withBoundRoomContext(
-    ...args: Parameters<typeof useHistoryVersionData>
+  function useHistoryVersionYjsData_withBoundRoomContext(
+    ...args: Parameters<typeof useHistoryVersionYjsData>
   ) {
-    return useHistoryVersionData_withRoomContext(BoundRoomContext, ...args);
+    return useHistoryVersionYjsData_withRoomContext(BoundRoomContext, ...args);
   }
 
   function useRoomSubscriptionSettings_withBoundRoomContext() {
@@ -4299,7 +4313,9 @@ export function createRoomContext<
     // prettier-ignore
     useHistoryVersions: useHistoryVersions_withBoundRoomContext as TRoomBundle["useHistoryVersions"],
     // prettier-ignore
-    useHistoryVersionData: useHistoryVersionData_withBoundRoomContext as TRoomBundle["useHistoryVersionData"],
+    useHistoryVersionData: useHistoryVersionYjsData_withBoundRoomContext as TRoomBundle["useHistoryVersionData"],
+    // prettier-ignore
+    useHistoryVersionYjsData: useHistoryVersionYjsData_withBoundRoomContext as TRoomBundle["useHistoryVersionYjsData"],
 
     // prettier-ignore
     useRoomSubscriptionSettings: useRoomSubscriptionSettings_withBoundRoomContext as TRoomBundle["useRoomSubscriptionSettings"],
@@ -5076,6 +5092,7 @@ export {
   useHistoryVersionData,
   _useHistoryVersions as useHistoryVersions,
   _useHistoryVersionsSuspense as useHistoryVersionsSuspense,
+  useHistoryVersionYjsData,
   _useIsInsideRoom as useIsInsideRoom,
   useLostConnectionListener,
   useMarkRoomThreadAsRead,